resolved another dep cycle
